refactor: retire font services
