+++ pyproject.toml
@@ -130,6 +130,7 @@ files = [
     "mpvqc/appearance/models/color_scheme_preference.py",
     "mpvqc/appearance/services/__init__.py",
     "mpvqc/appearance/services/color_scheme.py",
+    "mpvqc/appearance/services/fonts.py",
     "mpvqc/appearance/services/palette_catalog.py",
     "mpvqc/appearance/services/preferences.py",
     "mpvqc/appearance/services/schemes.py",
@@ -249,8 +250,6 @@ files = [
     "mpvqc/services/application_paths.py",
     "mpvqc/services/desktop.py",
     "mpvqc/services/file_startup.py",
-    "mpvqc/services/font_loader.py",
-    "mpvqc/services/label_width_calculator.py",
     "mpvqc/services/resource.py",
     "mpvqc/services/settings.py",
     "mpvqc/services/settings_file.py",
